refactor: remove use of dynamic type in the Validator class by introducing an IValidatorRule interface
The abstract Validator class will use the IValidatorRule interface to store rules. This will allow the Validator to access the rules' IsValid method. It does not need to be concenerd with the methods and properties in the internal IPropertyRule interface. Internal classes like RuleBuilder will use the IPropertyRule which implements the IValidatorRule and will therefore have access to stored specifications, cascade mode etc.